requirements.txt
# Pulse Dependencies

# Core dependencies
numpy>=1.24.0
pydantic>=2.0.0

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-flask>=0.41b0
opentelemetry-instrumentation-django>=0.41b0

# Storage
redis>=4.5.0
msgpack>=1.0.0
elasticsearch>=8.0.0

# ML (optional but recommended)
scikit-learn>=1.3.0
torch>=2.0.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
starlette>=0.27.0

# HTTP client
httpx>=0.24.0
aiohttp>=3.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
